[OP-2] fix: refactor/rename dataset loader
